runpod-deployment/source-code/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart
pyyaml
openai>=1.0.0
numpy>=1.21.0
aiofiles
aiohttp>=3.8.0

# HiRAG dependencies (from HiRAG/requirements.txt)
graspologic==3.4.1
jsonlines==4.0.0
matplotlib>=3.5.0
nano_vectordb==0.0.2
neo4j==5.25.0
nest_asyncio==1.6.0
networkx>=2.8,<3.3
pandas>=1.3.0
ragas==0.2.2
scikit-learn>=1.0.0
seaborn==0.13.2
setuptools==73.0.1
tenacity==9.0.0
tiktoken==0.7.0
tqdm==4.66.5
transformers>=4.20.0
umap_learn==0.5.6
xxhash==3.5.0
future==1.0.0
spire.doc
spire.presentation
PyMuPDF>=1.20.0
docx2pdf